fix(lint): master is red — unnecessary uint64 conversion in inodeOf (#346)
`golangci-lint` fails on internal/api/island_secrets_mount_test.go:24
(unconvert), so the lint job on master has been red since 550ca9a and
every open PR inherits it.

syscall.Stat_t.Ino is already uint64 on both platforms CI builds, so the
conversion is a no-op. Dropping it rather than adding a nolint: there is
nothing here worth suppressing a linter over.

Checked that this does not trade a lint failure for a build failure on
the other platform — `GOOS=darwin go vet ./internal/api/` typechecks the
test files and passes, as does the linux test run and the full suite.

+++ internal/api/island_secrets_mount_test.go
@@ -21,7 +21,7 @@ func inodeOf(t *testing.T, path string) uint64 {
 	if err := syscall.Stat(path, &st); err != nil {
 		t.Fatalf("stat %s: %v", path, err)
 	}
-	return uint64(st.Ino)
+	return st.Ino
 }
 
 // What `dejima secret set` and `secret rm` used to do on a running island:
